refactor: extract overlays/shared/CompoundNode
