+++ packages/diagram/src/overlays/edge-details/EdgeDetailsXYFlow.tsx
@@ -15,7 +15,7 @@ import { useDiagramStoreApi } from '../../hooks/useDiagramState'
 import { SelectEdge } from './SelectEdge'
 import * as css from './SelectEdge.css'
 import { useLayoutedEdgeDetails, ZIndexes } from './use-layouted-edge-details'
-import { CompoundNode } from './xyflow/CompoundNode'
+import { CompoundNode } from '../shared/xyflow/CompoundNode'
 import { ElementNode } from './xyflow/ElementNode'
 import { RelationshipEdge } from './xyflow/RelationshipEdge'
 import type { BaseTypes } from '../shared/_types'
