+++ packages/diagram/src/overlays/relationships-of/RelationshipsXYFlow.tsx
@@ -21,7 +21,7 @@ import { centerXYInternalNode } from '../../xyflow/utils'
 import { cssReactflowMarker } from '../Overlays.css'
 import { type RelationshipsOfTypes } from './_types'
 import { ZIndexes } from './use-layouted-relationships'
-import { CompoundNode } from './xyflow/CompoundNode'
+import { CompoundNode } from '../shared/xyflow/CompoundNode'
 import { ElementNode } from './xyflow/ElementNode'
 import { EmptyNode } from './xyflow/EmptyNode'
 import { RelationshipEdge } from './xyflow/RelationshipEdge'
